pyproject.toml
[tool.poetry]
name = "pmm-cfg-gen"
version = "0.2.3"
description = "A script to help automatically generate Plex Meta Manager configuration files for your libraries"
license = "GPL-3.0-or-later"
authors = ["Shawn Anderson <sanderson@eye-catcher.com>"]
readme = "README.md"
homepage = "https://github.com/ravensorb/Plex-Meta-Manager-Config-Generator"
repository = "https://github.com/ravensorb/Plex-Meta-Manager-Config-Generator"
keywords = [ "pmm", "plex-meta-manager", "plex" ]
classifiers = [
    "Topic :: Utilities"
]
packages = [
    {include = "pmm_cfg_gen", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
jsonpickle = "*"
plexapi = "*"
jinja2 = "*"
ruamel-yaml = "^0.17.21"
expandvars = "^0.9.0"
python-dotenv = "^1.0.0"
confuse = "^2.0.1"
click = "^8.1.3"
importlib-resources = "^5.12.0"
coloredlogs = "^15.0.1"
readchar = "^4.0.5"
timedelta = "^2020.12.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
scriv = {extras = ["toml"], version = "^1.2.1"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "*"

[tool.poetry.scripts]
"pmm-cfg-gen" = "pmm_cfg_gen:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
